converting the download manager to ObjC

+++ Pod/Classes/TCBlobDownload.h
@@ -22,7 +22,7 @@
  :param: totalBytesWritten The total number of bytes the download has currently written to the disk.
  :param: totalBytesExpectedToWrite The total number of bytes the download will write to the disk once completed.
  */
--(void)download:(TCBlobDownload *)download didProgress:(CGFloat)progress totalBytesWritten:(NSInteger)totalBytesWritten totalBytesExpectedToWrite:(NSInteger)totalBytesExpectedToWrite;
+-(void)download:(TCBlobDownload *)download didProgress:(double)progress totalBytesWritten:(int64_t)totalBytesWritten totalBytesExpectedToWrite:(int64_t)totalBytesExpectedToWrite;
 
 /**
  Informs the delegate that the download was completed (similar to `NSURLSession -URLSession:task:didCompleteWithError:`).
@@ -36,7 +36,7 @@
 -(void)download:(TCBlobDownload *)download didFinishWithError:(NSError *)error atLocation:(NSURL *)location;
 @end
 
-typedef void (^TCBlobDownloadProgressHandler)(CGFloat progress, NSInteger totalBytesWritten, NSInteger totalBytesExpectedToWrite);
+typedef void (^TCBlobDownloadProgressHandler)(double progress, int64_t totalBytesWritten, int64_t totalBytesExpectedToWrite);
 typedef void (^TCBlobDownloadCompletionHandler)(NSError *error, NSURL *location);
 
 @interface TCBlobDownload : NSObject
@@ -53,7 +53,7 @@ typedef void (^TCBlobDownloadCompletionHandler)(NSError *error, NSURL *location)
 @property (nonatomic, strong) NSURL *directory;
 @property (nonatomic, strong) NSError *error;
 
-@property (nonatomic, assign) CGFloat progress;
+@property (nonatomic, assign) double progress;
 
 @property (nonatomic, copy) NSString *fileName;
 @property (nonatomic, strong) NSURL *destinationURL;
